JBIDE-14548 fix IAE while installing plugins via project examples requirements
- Moved mylyn discovery from central to project-examples
- refactored 3 ConnectorDiscovery creation points into one DiscoveryUtil
  method

Author:    Denis Golovin <dgolovin@gmail.com>
CoAuthor:    Fred Bricon <fbricon@gmail.com>
