+++ central/plugins/org.jboss.tools.central/META-INF/MANIFEST.MF
@@ -38,10 +38,7 @@ Export-Package: org.jboss.tools.central,
  org.jboss.tools.central.editors,
  org.jboss.tools.central.editors.xpl,
  org.jboss.tools.central.internal;x-friends:="org.jboss.tools.central.test",
- org.jboss.tools.central.internal.discovery;x-friends:="org.jboss.tools.central.test",
  org.jboss.tools.central.internal.discovery.wizards;x-friends:="org.jboss.tools.central.test",
- org.jboss.tools.central.internal.discovery.wizards.xpl;x-friends:="org.jboss.tools.central.test",
- org.jboss.tools.central.internal.xpl,
  org.jboss.tools.central.jobs,
  org.jboss.tools.central.model,
  org.jboss.tools.central.preferences
