+++ central/plugins/org.jboss.tools.central/src/org/jboss/tools/central/editors/xpl/DiscoveryViewer.java
@@ -122,10 +122,12 @@ import org.eclipse.ui.forms.IFormColors;
 import org.eclipse.ui.progress.WorkbenchJob;
 import org.eclipse.ui.statushandlers.StatusManager;
 import org.eclipse.ui.themes.IThemeManager;
-import org.jboss.tools.central.internal.discovery.ChainedDiscoveryStrategy;
-import org.jboss.tools.central.internal.discovery.ChainedDiscoveryStrategy.DataCollector;
-import org.jboss.tools.central.internal.discovery.ExpressionBasedBundleDiscoveryStrategy;
-import org.jboss.tools.central.internal.discovery.ExpressionBasedRemoteBundleDiscoveryStrategy;
+import org.jboss.tools.project.examples.internal.discovery.ChainedDiscoveryStrategy;
+import org.jboss.tools.project.examples.internal.discovery.ChainedDiscoveryStrategy.DataCollector;
+import org.jboss.tools.project.examples.internal.discovery.ChainedDiscoveryStrategy.DiscoveryConnectorCollector;
+import org.jboss.tools.project.examples.internal.discovery.DiscoveryUtil;
+import org.jboss.tools.project.examples.internal.discovery.ExpressionBasedBundleDiscoveryStrategy;
+import org.jboss.tools.project.examples.internal.discovery.ExpressionBasedRemoteBundleDiscoveryStrategy;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.Version;
 
@@ -137,8 +139,6 @@ import org.osgi.framework.Version;
  */
 public class DiscoveryViewer {
 
-	private static boolean ALLOW_DUPLICATE_DISCOVERY_CONNECTORS = Boolean.getBoolean("org.jboss.tools.central.allow.duplicate.connectors");
-	
 	public class ConnectorBorderPaintListener implements PaintListener {
 		public void paintControl(PaintEvent e) {
 			Composite composite = (Composite) e.widget;
@@ -1493,24 +1493,7 @@ public class DiscoveryViewer {
 					if (DiscoveryViewer.this.installedFeatures == null) {
 						DiscoveryViewer.this.installedFeatures = getInstalledFeatures(monitor);
 					}
-
-					ConnectorDiscovery connectorDiscovery = new ConnectorDiscovery();
-					ChainedDiscoveryStrategy chainedDiscoveryStrategy = new ChainedDiscoveryStrategy(new DiscoveryConnectorCollector(ALLOW_DUPLICATE_DISCOVERY_CONNECTORS));
-
-					// look for remote descriptor first
-					if (directoryUrl != null) {
-						ExpressionBasedRemoteBundleDiscoveryStrategy remoteDiscoveryStrategy = new ExpressionBasedRemoteBundleDiscoveryStrategy();
-						remoteDiscoveryStrategy.setDirectoryUrl(directoryUrl);
-						chainedDiscoveryStrategy.addStrategy(remoteDiscoveryStrategy);
-					}
-
-					//Fall back on bundle discovery strategy
-					chainedDiscoveryStrategy.addStrategy(new ExpressionBasedBundleDiscoveryStrategy());
-
-					// look for descriptors from installed bundles
-					connectorDiscovery.getDiscoveryStrategies().add(chainedDiscoveryStrategy);
-
-
+					ConnectorDiscovery connectorDiscovery = DiscoveryUtil.createConnectorDiscovery(directoryUrl);
 					connectorDiscovery.setEnvironment(environment);
 					connectorDiscovery.setVerifyUpdateSiteAvailability(false);
 					try {
@@ -1600,28 +1583,6 @@ public class DiscoveryViewer {
 
 	protected Set<String> getInstalledFeatures(IProgressMonitor monitor) throws InterruptedException {
 		return DiscoveryUi.createInstallJob().getInstalledFeatures(monitor);
-	}
-
-	private final class DiscoveryConnectorCollector implements DataCollector {
-		
-		private boolean isComplete;
-		
-		private boolean allowDuplicates;
-		
-		public DiscoveryConnectorCollector(boolean allowDuplicates) {
-			this.allowDuplicates = allowDuplicates;
-		}
-		
-		@Override
-		public boolean isComplete() {
-			return isComplete && !allowDuplicates;
-		}
-
-		@Override
-		public void collectData(AbstractDiscoveryStrategy ds) {
-			List<DiscoveryConnector> collected = ds.getConnectors();
-			isComplete = (collected != null && !collected.isEmpty());
-		}		
 	}
 
 }
