+++ central/plugins/org.jboss.tools.central/src/org/jboss/tools/central/internal/discovery/wizards/CachedProxyWizardDiscoveryStrategy.java
@@ -20,7 +20,7 @@ import org.eclipse.core.runtime.IExtension;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.mylyn.internal.discovery.core.model.AbstractDiscoverySource;
 import org.jboss.tools.central.JBossCentralActivator;
-import org.jboss.tools.central.internal.discovery.wizards.xpl.ExternalBundleDiscoveryStrategy;
+import org.jboss.tools.project.examples.internal.discovery.xpl.ExternalBundleDiscoveryStrategy;
 
 /**
  * @author Fred Bricon
