+++ central/plugins/org.jboss.tools.central/src/org/jboss/tools/central/internal/discovery/wizards/InstalledProxyWizardDiscoveryStrategy.java
@@ -19,7 +19,7 @@ import org.eclipse.core.runtime.IContributor;
 import org.eclipse.core.runtime.IExtension;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.mylyn.internal.discovery.core.model.AbstractDiscoverySource;
-import org.jboss.tools.central.internal.discovery.wizards.xpl.BundleDiscoveryStrategy;
+import org.jboss.tools.project.examples.internal.discovery.xpl.BundleDiscoveryStrategy;
 
 /**
  * @author Fred Bricon
