+++ central/tests/org.jboss.tools.central.test/src/org/jboss/tools/central/internal/discovery/wizards/ChainedDiscoveryStrategyTest.java
@@ -25,7 +25,7 @@ import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.MultiStatus;
 import org.eclipse.core.runtime.Status;
 import org.eclipse.mylyn.internal.discovery.core.model.AbstractDiscoveryStrategy;
-import org.jboss.tools.central.internal.discovery.ChainedDiscoveryStrategy;
+import org.jboss.tools.project.examples.internal.discovery.ChainedDiscoveryStrategy;
 import org.junit.Test;
 
 public class ChainedDiscoveryStrategyTest {
