+++ examples/plugins/org.jboss.tools.project.examples/META-INF/MANIFEST.MF
@@ -51,6 +51,9 @@ Export-Package: org.jboss.tools.project.examples,
  org.jboss.tools.project.examples.dialog,
  org.jboss.tools.project.examples.dialog.xpl,
  org.jboss.tools.project.examples.fixes,
+ org.jboss.tools.project.examples.internal.discovery;x-friends:="org.jboss.tools.central",
+ org.jboss.tools.project.examples.internal.discovery.xpl;x-internal:=true,
+ org.jboss.tools.project.examples.internal.xpl;x-internal:=true,
  org.jboss.tools.project.examples.model,
  org.jboss.tools.project.examples.preferences,
  org.jboss.tools.project.examples.runtimes,
