+++ examples/plugins/org.jboss.tools.project.examples/src/org/jboss/tools/project/examples/internal/discovery/ExpressionBasedBundleDiscoveryStrategy.java
@@ -8,28 +8,28 @@
  * Contributors:
  *     JBoss by Red Hat - Initial implementation.
  ************************************************************************************/
-package org.jboss.tools.central.internal.discovery;
+package org.jboss.tools.project.examples.internal.discovery;
 
 import org.eclipse.core.runtime.IContributor;
 import org.eclipse.core.runtime.IExtension;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.mylyn.internal.discovery.core.model.AbstractDiscoverySource;
-import org.eclipse.mylyn.internal.discovery.core.model.RemoteBundleDiscoveryStrategy;
+import org.eclipse.mylyn.internal.discovery.core.model.BundleDiscoveryStrategy;
 
 /**
- * A subclass of RemoteBundleDiscoveryStrategy to avoid validation fails when property expressions are used in connector description 
+ * A subclass of BundleDiscoveryStrategy to avoid validation fails when property expressions are used in connector description 
  * this class only works when {@link ExpressionBasedDiscoveryConnector} is used when loading the connector descriptor
  * 
- * @see org.eclipse.mylyn.internal.discovery.core.model.RemoteBundleDiscoveryStrategy
+ * @see org.eclipse.mylyn.internal.discovery.core.model.BundleDiscoveryStrategy
  * @author snjeza
  */
-public class ExpressionBasedRemoteBundleDiscoveryStrategy extends RemoteBundleDiscoveryStrategy {
+public class ExpressionBasedBundleDiscoveryStrategy extends BundleDiscoveryStrategy {
 
 	private ExpressionBasedDiscoveryExtensionProcessor processor = new ExpressionBasedDiscoveryExtensionProcessor() {
 		@Override
 		public AbstractDiscoverySource computeDiscoverySource(
 				IContributor contributor) {
-			return ExpressionBasedRemoteBundleDiscoveryStrategy.this.computeDiscoverySource(contributor);
+			return ExpressionBasedBundleDiscoveryStrategy.this.computeDiscoverySource(contributor);
 		}
 	};
