+++ examples/plugins/org.jboss.tools.project.examples/src/org/jboss/tools/project/examples/internal/discovery/xpl/DiscoveryRegistryStrategy.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     Tasktop Technologies - initial API and implementation
  *******************************************************************************/
-package org.jboss.tools.central.internal.discovery.wizards.xpl;
+package org.jboss.tools.project.examples.internal.discovery.xpl;
 
 import java.io.File;
 import java.io.IOException;
